+++ hdl/en_cl_fix_pkg.vhd
@@ -1,5 +1,5 @@
 ---------------------------------------------------------------------------------------------------
--- Copyright (c) 2024 Enclustra GmbH, Switzerland (info@enclustra.com)
+-- Copyright (c) 2025 Enclustra GmbH, Switzerland (info@enclustra.com)
 -- 
 -- Permission is hereby granted, free of charge, to any person obtaining a copy of this software
 -- and associated documentation files (the "Software"), to deal in the Software without
@@ -63,6 +63,13 @@ package en_cl_fix_pkg is
         Warn_s,         -- No saturation, only warning.
         Sat_s,          -- Only saturation, no warning.
         SatWarn_s       -- Saturation and warning.
+    );
+    
+    type RegisterMode_t is
+    (
+        Auto_s,         -- Inserts the recommended registering. See cl_fix_recommended_pipelining.
+        Yes_s,          -- Inserts all registering. Can be useful for consistent latency.
+        No_s            -- Inserts no registering. Use with caution (poor timing performance).
     );
     
     -----------------------------------------------------------------------------------------------
@@ -153,6 +160,29 @@ package en_cl_fix_pkg is
         result_fmt  : FixFormat_t;
         round       : FixRound_t := Trunc_s
     ) return boolean;
+    
+    -- Recommended pipeline stages for cl_fix_round
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        round       : FixRound_t;
+        fmt_check   : boolean := true
+    ) return natural;
+    
+    -- Recommended pipeline stages for cl_fix_saturate
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        saturate    : FixSaturate_t
+    ) return natural;
+    
+    -- Recommended pipeline stages for cl_fix_resize
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        round       : FixRound_t;
+        saturate    : FixSaturate_t
+    ) return natural;
     
     -----------------------------------------------------------------------------------------------
     -- Math Functions
@@ -298,7 +328,7 @@ package body en_cl_fix_pkg is
     
     function convert(a : std_logic_vector; aFmt, rFmt : FixFormat_t) return std_logic_vector is
         -- Force downto 0
-        constant a_c    : std_logic_vector(a'length-1 downto 0) := a;
+        constant a_c    : std_logic_vector(cl_fix_width(aFmt)-1 downto 0) := a;
         
         -- This function converts from aFmt to rFmt without any rounding or saturation:
         --     - It does *not* support rFmt.F < aFmt.F (offset_c is type natural). To reduce frac
@@ -768,6 +798,13 @@ package body en_cl_fix_pkg is
     end;
     
     function cl_fix_from_real(a : real; result_fmt : FixFormat_t; saturate : FixSaturate_t := SatWarn_s) return std_logic_vector is
+        -- Several toolchains have bugs in ieee.math_real."mod" (Vivado, Efinity, Gowin EDA).
+        -- Therefore, this local function is used as a workaround.
+        function real_mod(a, b : real) return real is
+        begin
+            return a - b * floor(a/b);
+        end function;
+        
         constant ChunkSize_c    : positive := 30;
         constant ChunkCount_c   : positive := (cl_fix_width(result_fmt) + ChunkSize_c - 1)/ChunkSize_c;
         variable ASat_v         : real;
@@ -794,13 +831,7 @@ package body en_cl_fix_pkg is
         
         -- Convert to fixed-point in chunks (required for formats that don't fit into integer)
         for i in 0 to ChunkCount_c-1 loop
-            -- Note: Due to a Xilinx Vivado bug, we must explicitly call the math_real mod operator
-            ChunkInt_v := integer(ieee.math_real."mod"(ASat_v, 2.0**ChunkSize_c));
-            if ChunkInt_v > 0 then
-                Chunk_v := std_logic_vector(to_unsigned(ChunkInt_v, ChunkSize_c));
-            else
-                Chunk_v := std_logic_vector(to_signed(ChunkInt_v, ChunkSize_c));
-            end if;
+            Chunk_v := std_logic_vector(to_unsigned(integer(real_mod(ASat_v, 2.0**ChunkSize_c)), ChunkSize_c));
             Result_v((i+1)*ChunkSize_c-1 downto i*ChunkSize_c) := Chunk_v;
             ASat_v := floor(ASat_v/2.0**ChunkSize_c);
         end loop;
@@ -885,7 +916,7 @@ package body en_cl_fix_pkg is
         fmt_check   : boolean := true
     ) return std_logic_vector is
         -- Force downto 0
-        constant a_c            : std_logic_vector(a'length-1 downto 0) := a;
+        constant a_c            : std_logic_vector(cl_fix_width(a_fmt)-1 downto 0) := a;
         
         -- The result format takes care of potential integer growth due to the rounding mode.
         -- In the intermediate calculation, we need to +/- up to 2.0**-(result_fmt.F+1) in order to
@@ -905,6 +936,7 @@ package body en_cl_fix_pkg is
         variable mid_v          : unsigned(cl_fix_width(mid_fmt_c)-1 downto 0) := (others => '0');
         variable result_v       : std_logic_vector(cl_fix_width(result_fmt)-1 downto 0);
     begin
+        -- Allow the designer to ignore the worst-case result format (with caution).
         if fmt_check then
             assert result_fmt = cl_fix_round_fmt(a_fmt, result_fmt.F, round)
                 report "cl_fix_round: Invalid result format. Use cl_fix_round_fmt()." severity Failure;
@@ -1005,6 +1037,76 @@ package body en_cl_fix_pkg is
     begin
         return cl_fix_compare(">=", Rounded_c, rndFmt_c, cl_fix_min_value(result_fmt), result_fmt) and
                cl_fix_compare("<=", Rounded_c, rndFmt_c, cl_fix_max_value(result_fmt), result_fmt);
+    end;
+    
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        round       : FixRound_t;
+        fmt_check   : boolean := true
+    ) return natural is
+    begin
+        -- Allow the designer to ignore the worst-case result format (with caution).
+        if fmt_check then
+            assert result_fmt = cl_fix_round_fmt(a_fmt, result_fmt.F, round)
+                report "cl_fix_recommended_pipelining: Invalid result format. Use cl_fix_round_fmt()." severity Failure;
+        end if;
+        
+        -- Registering is not needed if zero logic is used. This happens in two cases:
+        -- (1) During truncation.
+        if round = Trunc_s then
+            return 0;
+        else
+            -- If a new rounding mode is defined, then appropriate behavior must be implemented.
+            assert round = NonSymPos_s or round = NonSymNeg_s or round = SymInf_s or round = SymZero_s or round = ConvEven_s or round = ConvOdd_s
+                report "cl_fix_recommended_pipelining: Unhandled rounding mode."
+                severity Failure;
+        end if;
+        -- (2) If the number of fractional bits isn't being decreased.
+        if result_fmt.F >= a_fmt.F then
+            return 0;
+        end if;
+        return 1;
+    end;
+    
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        saturate    : FixSaturate_t
+    ) return natural is
+    begin
+        assert result_fmt.F = a_fmt.F
+            report "cl_fix_recommended_pipelining: Number of frac bits cannot change during saturation."
+            severity Failure;
+        
+        -- Registering is not needed if zero logic is used. This happens in two cases:
+        -- (1) During wrapping.
+        if saturate = None_s or saturate = Warn_s then
+            return 0;
+        else
+            -- If a new saturation mode is defined, then appropriate behavior must be implemented.
+            assert saturate = Sat_s or saturate = SatWarn_s
+                report "cl_fix_recommended_pipelining: Unhandled saturation mode."
+                severity Failure;
+        end if;
+        -- (2) If the number of integer bits is not being decreased, and the number of sign bits is
+        --     not being changed.
+        if result_fmt.I >= a_fmt.I and result_fmt.S = a_fmt.S then
+            return 0;
+        end if;
+        return 1;
+    end;
+    
+    function cl_fix_recommended_pipelining(
+        a_fmt       : FixFormat_t;
+        result_fmt  : FixFormat_t;
+        round       : FixRound_t;
+        saturate    : FixSaturate_t
+    ) return natural is
+        constant round_fmt_c    : FixFormat_t := cl_fix_round_fmt(a_fmt, result_fmt.F, round);
+    begin
+        return cl_fix_recommended_pipelining(a_fmt, round_fmt_c, round)
+             + cl_fix_recommended_pipelining(round_fmt_c, result_fmt, saturate);
     end;
     
     function cl_fix_abs(
@@ -1120,8 +1222,8 @@ package body en_cl_fix_pkg is
         saturate    : FixSaturate_t := Warn_s
     ) return std_logic_vector is
         -- Force downto 0
-        constant a_c            : std_logic_vector(a'length-1 downto 0) := a;
-        constant b_c            : std_logic_vector(b'length-1 downto 0) := b;
+        constant a_c            : std_logic_vector(cl_fix_width(a_fmt)-1 downto 0) := a;
+        constant b_c            : std_logic_vector(cl_fix_width(b_fmt)-1 downto 0) := b;
         
         -- VHDL doesn't define a * operator for mixed signed*unsigned or unsigned*signed.
         -- Just inside cl_fix_mult, it is safe to define them for local use.
@@ -1213,7 +1315,7 @@ package body en_cl_fix_pkg is
     
     function cl_fix_sign(a : std_logic_vector; aFmt : FixFormat_t) return std_logic is
         -- Force downto 0
-        constant a_c    : std_logic_vector(a'length-1 downto 0) := a;
+        constant a_c    : std_logic_vector(cl_fix_width(aFmt)-1 downto 0) := a;
     begin
         if aFmt.S = 0 or cl_fix_width(aFmt) = 0 then
             return '0';